# PERF: _form_blocks