```
from asv_bench.benchmarks.frame_ctor import *

self = FromArrays()

self.setup()

%timeit self.time_frame_from_arrays_int()
2.28 ms ± 67 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- master
1.82 ms ± 5.34 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
```

Orthogonal to #43142, one more optimization after these two are in.