@jreback 2021-08-21
https://github.com/pandas-dev/pandas/pull/43144/checks?check_run_id=3386832245 looks like a real error