@jbrockmendel 2021-08-21
OK, looks like in numpy<1.21 comparing np.dtype object to EA dtype object raises instead of returning False, so the itertools.groupby is failing.  will figure out a workaround