@jbrockmendel 2021-08-22
updated with np<1.21-compatiblity + green